Add ability to content script to fetch content by UUID
